Bug fix : resqml2::AbstractGridRepresentation::getParentCellIndexCount fails
Fix #352
IntervalCount, ParentCellIndexCount and ParentColumnIndexCount are now uint64_t in  resqml2::AbstractGridRepresentation
ElementCountPerValue is now uint64_t in  resqml2::AbstractProperty

+++ src/resqml2/ContinuousProperty.h
@@ -583,7 +583,7 @@ namespace RESQML2_NS
 			unsigned int numArrayDimensions,
 			T * minimumValue = nullptr, T * maximumValue = nullptr)
 		{
-			const unsigned int elementCount = getElementCountPerValue();
+			const uint64_t elementCount = getElementCountPerValue();
 
 			// Some minimum and maximum values are given : No need to compute them.
 			if (minimumValue != nullptr) {
